requirements.txt
hidapi
matplotlib
numpy
pandas
